resolve interference
