+++ src/main/java/weby/kiwi/domain/collection/dto/CollectionResDto.java
@@ -1,0 +1,19 @@
+package weby.kiwi.domain.collection.dto;
+
+public class CollectionResDto {
+    private int user_id;
+    private int year;
+    private int month;
+
+    public int getUser_id() {
+        return user_id;
+    }
+
+    public int getYear() {
+        return year;
+    }
+
+    public int getMonth() {
+        return month;
+    }
+}
